+++ src/components/CardService/index.scss
@@ -19,13 +19,11 @@
     .card-title{
         color: var(--text-white);
         margin-bottom: 15px;
-        font-size: 1.2em;
     }
     .card-description {
         color: var(--text-color-secondary);
         text-align: justify;
         line-height: var(--line-height-standar);
-        font-size: 1.1em;
     }
     .card-footer{
         margin-bottom: 20px;
