+++ src/components/Clients/index.js
@@ -1,20 +1,37 @@
 import CardClient from "./CardClient";
+import "./index.scss";
 
 const listClients = [
   {
     img: "/assets/img/clients/logo-huella-digital.svg",
+    filterStyle: "img-filter-gray",
+    site: "https://huella-digital.mx",
   },
-  { img: "/assets/img/clients/logo-blanco-pwevents.png" },
-  { img: "/assets/img/clients/logo-marcshuttle-black.png" },
-  { img: "/assets/img/clients/logo-clinicadentalquirurjica.png" },
+  {
+    img: "/assets/img/clients/logo-pwevents.png",
+    site: "https://pweventscancun.com",
+  },
+  {
+    img: "/assets/img/clients/logo-marcshuttle-black.png",
+    filterStyle: "img-filter-invert",
+    site: "https://marcshuttlecancun.com",
+  },
+  {
+    img: "/assets/img/clients/logo-clinicadentalquirurjica.png",
+    filterStyle: "img-filter-gray",
+    site: "https://clinicadentalquirurgica.com",
+  },
 ];
 
 export const Clients = () => {
   return (
-    <div className="d-flex">
-      {listClients.map((item, index) => (
-        <CardClient key={index} srcImg={item.img} />
-      ))}
+    <div className="clients">
+      <h3 className="t-c-white t-size-l mb-3">Nuestros clientes</h3>
+      <div className="content-clients">
+        {listClients.map((item, index) => (
+          <CardClient key={index} dataClient={item} />
+        ))}
+      </div>
     </div>
   );
 };
